+++ threedi_models_and_simulations/data_models/simulation_data_models.py
@@ -85,6 +85,7 @@ class InitialConditions(SimulationElement):
     global_value_1d: float = None
     from_spatialite_1d: bool = None
     initial_waterlevels_1d: dict = None
+    online_waterlevels_1d: InitialWaterlevel = None
     global_value_2d: float = None
     online_raster_2d: InitialWaterlevel = None
     local_raster_2d: str = None